Rename P2P miscellaneous utils to just misc
