+++ elements/lisk-p2p/src/utils/index.ts
@@ -12,7 +12,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-export * from './miscellaneous';
+export * from './misc';
 export * from './sanitize';
 export * from './select';
 export * from './validate';
