+++ elements/lisk-p2p/src/utils/misc.ts
